Split up builddep
